@@ -1,53 +1,53 @@
-import XCTest
+import Testing
 
 @testable import App
 
-final class AsciidocTests: XCTestCase {
+@Suite struct AsciidocTests {
 
-  func testNonEntity160RomanNumeraled() {
-    XCTAssertEqual(
-      Asciidoc.htmlShortTitle("Epistles 133 &#8212; 160"),
-      "Epistles CXXXIII &#8212; CLX",
+  @Test func `non-entity 160 roman numeraled`() {
+    #expect(
+      Asciidoc.htmlShortTitle("Epistles 133 &#8212; 160")
+        == "Epistles CXXXIII &#8212; CLX",
     )
   }
 
-  func testHtmlTitleTurnsDoubleDashIntoEmdashEntity() throws {
-    XCTAssertEqual(Asciidoc.htmlTitle("Foo -- bar"), "Foo &#8212; bar")
+  @Test func `html title turns double dash into emdash entity`() {
+    #expect(Asciidoc.htmlTitle("Foo -- bar") == "Foo &#8212; bar")
   }
 
-  func testHtmlTitleChangesTrailingDigitsIntoRomanNumerals() throws {
-    XCTAssertEqual(Asciidoc.htmlTitle("Foo 3"), "Foo III")
+  @Test func `html title changes trailing digits into roman numerals`() {
+    #expect(Asciidoc.htmlTitle("Foo 3") == "Foo III")
   }
 
-  func testHtmlTitleDoesNotChangeYearsToRoman() throws {
-    XCTAssertEqual(
-      Asciidoc.htmlTitle("Chapter 9. Letters from 1818--1820"),
-      "Chapter IX. Letters from 1818&#8212;1820",
+  @Test func `html title does not change years to roman`() {
+    #expect(
+      Asciidoc.htmlTitle("Chapter 9. Letters from 1818--1820")
+        == "Chapter IX. Letters from 1818&#8212;1820",
     )
   }
 
-  func testHtmlShortTitleShortensVolumeToVol() throws {
-    XCTAssertEqual(
-      Asciidoc.htmlShortTitle("Foo -- Volume 1"),
-      "Foo &#8212; Vol.&#160;I",
+  @Test func `html short title shortens volume to vol`() {
+    #expect(
+      Asciidoc.htmlShortTitle("Foo -- Volume 1")
+        == "Foo &#8212; Vol.&#160;I",
     )
   }
 
-  func testHtmlShortTitleShortensSpanishVolumeToVol() throws {
-    XCTAssertEqual(
-      Asciidoc.htmlShortTitle("Foo -- volumen 4"),
-      "Foo &#8212; Vol.&#160;IV",
+  @Test func `html short title shortens spanish volume to vol`() {
+    #expect(
+      Asciidoc.htmlShortTitle("Foo -- volumen 4")
+        == "Foo &#8212; Vol.&#160;IV",
     )
   }
 
-  func testUtf8ShortTitleShortensCorrectly() throws {
-    XCTAssertEqual(
-      Asciidoc.utf8ShortTitle("Chapter 9. Letters from 1818--1820"),
-      "Chapter IX. Letters from 1818–1820",
+  @Test func `utf8 short title shortens correctly`() {
+    #expect(
+      Asciidoc.utf8ShortTitle("Chapter 9. Letters from 1818--1820")
+        == "Chapter IX. Letters from 1818–1820",
     )
   }
 
-  func testTrimmedUtf8ShortDocumentTitle() throws {
+  @Test func `trimmed utf8 short document title`() {
     let cases: [(String, Lang, String)] = [
       ("The Foobar", .en, "Foobar"),
       ("A Foobar", .en, "Foobar"),
@@ -61,7 +61,7 @@ final class AsciidocTests: XCTestCase {
     ]
 
     for (input, lang, expected) in cases {
-      XCTAssertEqual(Asciidoc.trimmedUtf8ShortDocumentTitle(input, lang: lang), expected)
+      #expect(Asciidoc.trimmedUtf8ShortDocumentTitle(input, lang: lang) == expected)
     }
   }
 }
