@@ -1,87 +1,86 @@
-import XCTest
-import XExpect
+import Testing
 
 @testable import App
 
-final class AudioPartValidityTests: XCTestCase {
-  override static func setUp() {
+@Suite struct AudioPartValidityTests {
+  init() {
     Current.logger = .null
   }
 
-  func testEmptyTitleInvalid() async {
+  @Test func `empty title invalid`() async {
     var part = AudioPart.valid
     part.title = ""
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testTooSmallDurationInvalid() async {
+  @Test func `too small duration invalid`() async {
     var part = AudioPart.valid
     part.duration = 120
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testTooSmallDurationValidIfNotPublished() async {
+  @Test func `too small duration valid if not published`() async {
     var part = AudioPart.valid
     part.mp3SizeLq = 0
     part.mp3SizeHq = 0
     part.duration = 120
-    await expect(part.isValid()).toBeTrue()
+    #expect(await part.isValid() == true)
   }
 
-  func testNegativeValuesForMp3SizeInvalid() async {
+  @Test func `negative values for mp3 size invalid`() async {
     var part = AudioPart.valid
     part.mp3SizeHq = -1
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
     part = AudioPart.valid
     part.mp3SizeLq = -1
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testSameSizeForMp3HqLqInvalid() async {
+  @Test func `same size for mp3 hq lq invalid`() async {
     var part = AudioPart.valid
     part.mp3SizeHq = 888_888
     part.mp3SizeLq = 888_888
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testLqMp3LargerThanHqInvalid() async {
+  @Test func `lq mp3 larger than hq invalid`() async {
     var part = AudioPart.valid
     part.mp3SizeHq = 888_888
     part.mp3SizeLq = 999_999
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testOrderLessThan1Iinvalid() async {
+  @Test func `order less than 1 invalid`() async {
     var part = AudioPart.valid
     part.order = 0
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testNegativeChapterValueInvalid() async {
+  @Test func `negative chapter value invalid`() async {
     var part = AudioPart.valid
     part.chapters = .init(-5)
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testNonSequencedChaptersInvalid() async {
+  @Test func `non-sequenced chapters invalid`() async {
     var part = AudioPart.valid
     part.chapters = .init(3, 5)
-    await expect(part.isValid()).toBeFalse()
+    #expect(await part.isValid() == false)
   }
 
-  func testZeroDraftStateItemsValid() async {
+  @Test func `zero draft state items valid`() async {
     var part = AudioPart.valid
     part.mp3SizeHq = 0
     part.mp3SizeLq = 0
-    await expect(part.isValid()).toBeTrue()
+    #expect(await part.isValid() == true)
   }
 
-  func testTempNoteToListenerAllowedWithSmallerSize() async {
+  @Test func `temp note to listener allowed with smaller size`() async {
     var part = AudioPart.valid
     part.title = "Nota para el oyente"
     part.mp3SizeLq = 300_000
     part.mp3SizeHq = 400_000
     part.duration = 25
-    await expect(part.isValid()).toBeTrue()
+    #expect(await part.isValid() == true)
   }
 }
