@@ -1,31 +1,31 @@
+import Testing
 import XCore
-import XCTest
 
-final class StringTests: XCTestCase {
-  func testSnakeCased() throws {
-    XCTAssertEqual("fooBarBaz".snakeCased, "foo_bar_baz")
+@Suite struct StringTests {
+  @Test func `snake cased`() {
+    #expect("fooBarBaz".snakeCased == "foo_bar_baz")
   }
 
-  func testShoutyCased() throws {
-    XCTAssertEqual("fooBarBaz".shoutyCased, "FOO_BAR_BAZ")
+  @Test func `shouty cased`() {
+    #expect("fooBarBaz".shoutyCased == "FOO_BAR_BAZ")
   }
 
-  func testRegexReplace() {
-    XCTAssertEqual("foobar".regexReplace("^foo", "jim"), "jimbar")
+  @Test func `regex replace`() {
+    #expect("foobar".regexReplace("^foo", "jim") == "jimbar")
   }
 
-  func testRegexRemove() {
-    XCTAssertEqual("foobar".regexRemove("^foo"), "bar")
+  @Test func `regex remove`() {
+    #expect("foobar".regexRemove("^foo") == "bar")
   }
 
-  func testMatchesRegex() {
-    XCTAssertEqual("foobar".matchesRegex("^foo"), true)
-    XCTAssertEqual("foobar".matchesRegex("bar$"), true)
-    XCTAssertEqual("foobar".matchesRegex("^bar"), false)
+  @Test func `matches regex`() {
+    #expect("foobar".matchesRegex("^foo") == true)
+    #expect("foobar".matchesRegex("bar$") == true)
+    #expect("foobar".matchesRegex("^bar") == false)
   }
 
-  func testPadLeft() {
-    XCTAssertEqual("foo".padLeft(toLength: 6, withPad: "_"), "___foo")
-    XCTAssertEqual("foo".padLeft(toLength: 8, withPad: "*"), "*****foo")
+  @Test func `pad left`() {
+    #expect("foo".padLeft(toLength: 6, withPad: "_") == "___foo")
+    #expect("foo".padLeft(toLength: 8, withPad: "*") == "*****foo")
   }
 }
